correct an error

+++ pointpats/quadrat_statistics.py
@@ -252,12 +252,11 @@ class HexagonM:
             possible_y_index_odd = int((point[1] - y_min) / (
                 self.semi_height * 2))
             if intercept_degree_x % 3 != 1:
-                center_index_x = (intercept_degree_x+1)/3
+                center_index_x = int((intercept_degree_x+1)/3)
                 center_index_y = possible_y_index_odd
                 if center_index_x % 2 == 0:
                     center_index_y = possible_y_index_even
-                dict_id_count[center_index_x + center_index_y *
-                              self.count_column] += 1
+                dict_id_count[center_index_x + center_index_y * self.count_column] += 1
             else: # two columns of cells can be possible
                 center_index_x = intercept_degree_x/3
                 center_x = center_index_x*semi_cell_length*3 + x_min
